requirements.txt
pandas==2.2.3
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
selenium==4.27.1
webdriver-manager==4.0.2
